wip: move status code to http interface
