+++ src/interface/http/result-to-http.test.ts
@@ -1,5 +1,5 @@
 import { createResultToHttpMapper, type ResultHttpMapping } from '@/interface/http/result-to-http';
-import { HTTP_CREATED, HTTP_INTERNAL_SERVER_ERROR, HTTP_UNAUTHORIZED } from '@/kairos/shared/http/status-code';
+import { HTTP_CREATED, HTTP_INTERNAL_SERVER_ERROR, HTTP_UNAUTHORIZED } from '@/interface/http/status-code';
 import { err } from '@/kairos/shared/result/err';
 import { ok } from '@/kairos/shared/result/ok';
 import { describe, expect, it } from 'vitest';
