+++ src/kairos/authentication/interface/http/login-response.ts
@@ -1,10 +1,10 @@
+import { type ResultHttpMapping } from '@/interface/http/result-to-http';
+import { HTTP_OK, HTTP_UNAUTHORIZED } from '@/interface/http/status-code';
 import {
   type AuthenticatedUser,
   authenticatedUserSerializerMetadata,
 } from '@/kairos/authentication/domain/authenticated-user';
 import { type InvalidCredentialsError } from '@/kairos/authentication/domain/errors';
-import { type ResultHttpMapping } from '@/interface/http/result-to-http';
-import { HTTP_OK, HTTP_UNAUTHORIZED } from '@/kairos/shared/http/status-code';
 
 export const loginResponse: ResultHttpMapping<AuthenticatedUser, InvalidCredentialsError> = {
   success: {
